+++ packages/components/elements/tabs/Tabs.tsx
@@ -1,54 +1,137 @@
 import * as React from "react";
 
-import { OrientationType } from "@_types/commonTypes";
 import * as TabsPrimitive from "@radix-ui/react-tabs";
-
 import { cn } from "@util/index";
+import { VariantProps, cva } from "class-variance-authority";
+import { tv } from "tailwind-variants";
+
+import { OrientationType } from "@_types/commonTypes";
+
 import { Chip, ChipTypes } from "../chip/Chip";
+
+const tabsListVariant = tv({
+  base: "",
+  variants: {
+    variant: {
+      default:
+        "hawa-flex hawa-w-fit hawa-flex-wrap hawa-items-center hawa-justify-start hawa-gap-1 hawa-rounded hawa-border hawa-bg-muted hawa-p-1 hawa-text-muted-foreground  dark:hawa-border-primary/10",
+      underlined:
+        "hawa-flex hawa-w-fit hawa-flex-wrap hawa-items-center hawa-justify-start hawa-gap-1 hawa-rounded  hawa-p-1 hawa-text-muted-foreground  dark:hawa-border-primary/10",
+      underlined_tabs:
+        "hawa-flex hawa-w-fit hawa-flex-wrap hawa-items-center hawa-justify-start hawa-gap-1 hawa-text-muted-foreground"
+    },
+    orientation: {
+      horizontal: "",
+      vertical: ""
+    }
+  },
+  compoundVariants: [
+    {
+      variant: "underlined_tabs",
+      orientation: "vertical",
+      class: "hawa-border-e-2 hawa-border-e-primary"
+    },
+    {
+      variant: "underlined_tabs",
+      orientation: "horizontal",
+      class: "hawa-border-b-2 hawa-border-b-primary"
+    }
+  ],
+  defaultVariants: {
+    variant: "default",
+    orientation: "horizontal"
+  }
+});
+
+const tabsTriggerVariant = tv({
+  base: "",
+  variants: {
+    variant: {
+      default:
+        "hawa-inline-flex hawa-w-full hawa-flex-1 hawa-select-none hawa-items-center hawa-justify-center hawa-gap-2 hawa-whitespace-nowrap hawa-rounded hawa-border hawa-px-3 hawa-py-1.5 hawa-text-sm hawa-font-medium hawa-ring-offset-background hawa-transition-all focus-visible:hawa-outline-none focus-visible:hawa-ring-2 focus-visible:hawa-ring-ring focus-visible:hawa-ring-offset-2 disabled:hawa-pointer-events-none disabled:hawa-opacity-50 data-[state=active]:hawa-bg-primary data-[state=active]:hawa-text-primary-foreground data-[state=active]:hawa-shadow-sm  dark:hawa-border-primary/10",
+      underlined:
+        "hawa-inline-flex  hawa-w-full hawa-flex-1 hawa-select-none hawa-items-center hawa-justify-center hawa-gap-2 hawa-whitespace-nowrap hawa-rounded hawa-rounded-none hawa-px-3 hawa-py-1.5 hawa-text-sm hawa-font-medium hawa-ring-offset-background hawa-transition-all focus-visible:hawa-outline-none focus-visible:hawa-ring-2 focus-visible:hawa-ring-ring focus-visible:hawa-ring-offset-2 disabled:hawa-pointer-events-none disabled:hawa-opacity-50",
+      underlined_tabs:
+        "hawa-inline-flex hawa-w-full hawa-flex-1 hawa-select-none hawa-items-center hawa-justify-center hawa-gap-2 hawa-whitespace-nowrap hawa-rounded   hawa-px-3 hawa-py-1.5 hawa-text-sm hawa-font-medium hawa-ring-offset-background hawa-transition-all focus-visible:hawa-outline-none focus-visible:hawa-ring-2 focus-visible:hawa-ring-ring focus-visible:hawa-ring-offset-2 disabled:hawa-pointer-events-none disabled:hawa-opacity-50 hawa-bg-primary/10 data-[state=active]:hawa-bg-primary data-[state=active]:hawa-text-primary-foreground  dark:hawa-border-primary/10"
+    },
+    orientation: {
+      horizontal: "",
+      vertical: ""
+    }
+  },
+  compoundVariants: [
+    {
+      variant: "underlined",
+      orientation: "horizontal",
+      class:
+        "data-[state=active]:hawa-border-b-primary hawa-border-b hawa-border-b-2"
+    },
+    {
+      variant: "underlined",
+      orientation: "vertical",
+      class:
+        "data-[state=active]:hawa-border-e-primary hawa-border-e hawa-border-e-2"
+    },
+    {
+      variant: "underlined_tabs",
+      orientation: "horizontal",
+      class: "hawa-rounded-b-none"
+    },
+    {
+      variant: "underlined_tabs",
+      orientation: "vertical",
+      class: "hawa-rounded-e-none"
+    }
+  ],
+  defaultVariants: {
+    variant: "default",
+    orientation: "horizontal"
+  }
+});
+
+type TabsVariants = "default" | "underlined" | "underlined_tabs";
 
 const TabsContext = React.createContext<{
   orientation?: OrientationType;
-}>({ orientation: "vertical" });
+  variant?: TabsVariants;
+}>({ orientation: "horizontal", variant: "default" });
 
 type TabsRootProps = React.ComponentPropsWithoutRef<
   typeof TabsPrimitive.Root
-> & { orientation?: "vertical" | "horizontal" };
+> & {
+  variant?: TabsVariants;
+};
 
 const Tabs = React.forwardRef<
   React.ElementRef<typeof TabsPrimitive.Root>,
   TabsRootProps
->(
-  (
-    { className, orientation = "vertical" as OrientationType, ...props },
-    ref
-  ) => (
-    <TabsPrimitive.Root
-      ref={ref}
-      className={cn(
-        "hawa-flex hawa-gap-2",
-        orientation === "vertical" ? "hawa-flex-row" : "hawa-flex-col",
-        className
-      )}
-      {...props}
-    >
-      <TabsContext.Provider value={{ orientation }}>
-        {props.children}
-      </TabsContext.Provider>
-    </TabsPrimitive.Root>
-  )
-);
+>(({ className, orientation, variant = "default", ...props }, ref) => (
+  <TabsPrimitive.Root
+    ref={ref}
+    className={cn(
+      "hawa-flex hawa-gap-2",
+      orientation === "vertical" ? "hawa-flex-row" : "hawa-flex-col",
+      className
+    )}
+    {...props}
+  >
+    <TabsContext.Provider value={{ orientation, variant }}>
+      {props.children}
+    </TabsContext.Provider>
+  </TabsPrimitive.Root>
+));
 Tabs.displayName = TabsPrimitive.Root.displayName;
 
 const TabsList = React.forwardRef<
   React.ElementRef<typeof TabsPrimitive.List>,
   React.ComponentPropsWithoutRef<typeof TabsPrimitive.List>
 >(({ className, ...props }, ref) => {
-  const { orientation } = React.useContext(TabsContext);
+  const { orientation, variant } = React.useContext(TabsContext);
   return (
     <TabsPrimitive.List
       ref={ref}
       className={cn(
-        "hawa-flex hawa-w-fit hawa-flex-wrap hawa-items-center hawa-justify-start hawa-gap-1 hawa-rounded hawa-border hawa-bg-muted hawa-p-1 hawa-text-muted-foreground  dark:hawa-border-primary/10",
+        tabsListVariant({ variant, orientation }),
         orientation === "vertical" ? "hawa-flex-col" : "hawa-flex-row",
         className
       )}
@@ -69,13 +152,11 @@ const TabsTrigger = React.forwardRef<
   React.ElementRef<typeof TabsPrimitive.Trigger>,
   TabsTriggerProps
 >(({ className, chipProps, ...props }, ref) => {
+  const { orientation, variant } = React.useContext(TabsContext);
   return (
     <TabsPrimitive.Trigger
       ref={ref}
-      className={cn(
-        "hawa-inline-flex hawa-w-full hawa-flex-1 hawa-select-none hawa-items-center hawa-justify-center hawa-gap-2 hawa-whitespace-nowrap hawa-rounded hawa-border hawa-px-3 hawa-py-1.5 hawa-text-sm hawa-font-medium hawa-ring-offset-background hawa-transition-all focus-visible:hawa-outline-none focus-visible:hawa-ring-2 focus-visible:hawa-ring-ring focus-visible:hawa-ring-offset-2 disabled:hawa-pointer-events-none disabled:hawa-opacity-50 data-[state=active]:hawa-bg-primary data-[state=active]:hawa-text-primary-foreground data-[state=active]:hawa-shadow-sm  dark:hawa-border-primary/10",
-        className
-      )}
+      className={cn(tabsTriggerVariant({ variant, orientation }), className)}
       {...props}
     >
       {props.children}