+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hawa-storybook
 
+## 0.26.40
+
+### Patch Changes
+
+- Updated dependencies
+  - @sikka/hawa@0.28.0
+
 ## 0.26.39
 
 ### Patch Changes